+++ system/modules/core/classes/FrontendTemplate.php
@@ -58,56 +58,6 @@ class FrontendTemplate extends \Template
 	{
 		global $objPage;
 
-		// Ignore certain URL parameters
-		$arrIgnore = array('id', 'file', 'token', 'day', 'month', 'year');
-
-		if ($GLOBALS['TL_CONFIG']['addLanguageToUrl'])
-		{
-			$arrIgnore[] = 'language';
-		}
-
-		$strParams = '';
-		$arrPageParams = array();
-
-		// Rebuild the URL to eliminate duplicate parameters
-		foreach (array_keys($_GET) as $key)
-		{
-			if (!in_array($key, $arrIgnore))
-			{
-				if (preg_match('/^page/', $key))
-				{
-					$arrPageParams[] = $key;
-					continue; // see #4141
-				}
-
-				if ($key == 'auto_item')
-				{
-					continue; // see #5482
-				}
-
-				if ($GLOBALS['TL_CONFIG']['useAutoItem'] && in_array($key, $GLOBALS['TL_AUTO_ITEM']))
-				{
-					$strParams .= '/' . \Input::get($key, false, true);
-				}
-				else
-				{
-					$strParams .= '/' . $key . '/' . \Input::get($key, false, true);
-				}
-			}
-		}
-
-		$strUrl = $this->generateFrontendUrl($objPage->row(), $strParams);
-
-		sort($arrPageParams); // see #4141
-		$strGlue = (!$GLOBALS['TL_CONFIG']['disableAlias'] && strpos($strUrl, '?') === false) ? '?' : '&';
-
-		// Re-add the page numbers
-		foreach ($arrPageParams as $key)
-		{
-			$strUrl .= $strGlue . $key . '=' . \Input::get($key);
-			$strGlue = '&';
-		}
-
 		$this->keywords = '';
 		$arrKeywords = array_map('trim', explode(',', $GLOBALS['TL_KEYWORDS']));
 
@@ -131,6 +81,7 @@ class FrontendTemplate extends \Template
 		}
 
 		$intCache = 0;
+		$strUrl = \Environment::get('request');
 
 		// Decide whether the page shall be cached
 		if (!isset($_GET['file']) && !isset($_GET['token']) && empty($_POST) && !BE_USER_LOGGED_IN && !FE_USER_LOGGED_IN && !$_SESSION['DISABLE_CACHE'] && !isset($_SESSION['LOGIN_ERROR']) && intval($objPage->cache) > 0 && !$objPage->protected)
@@ -225,18 +176,33 @@ class FrontendTemplate extends \Template
 			// Index protected pages if enabled
 			if ($GLOBALS['TL_CONFIG']['indexProtected'] || (!FE_USER_LOGGED_IN && !$objPage->protected))
 			{
-				$arrData = array
-				(
-					'url' => $strUrl,
-					'content' => $this->strBuffer,
-					'title' => $objPage->pageTitle ?: $objPage->title,
-					'protected' => ($objPage->protected ? '1' : ''),
-					'groups' => $objPage->groups,
-					'pid' => $objPage->id,
-					'language' => $objPage->language
-				);
+				$blnIndex = true;
 
-				\Search::indexPage($arrData);
+				// Do not index the page if certain parameters are set
+				foreach (array_keys($_GET) as $key)
+				{
+					if ($key == 'id' || $key == 'file' || $key == 'token' || $key == 'day' || $key == 'month' || $key == 'year' || $key == 'page' || strncmp($key, 'page_', 5) === 0)
+					{
+						$blnIndex = false;
+						break;
+					}
+				}
+
+				if ($blnIndex)
+				{
+					$arrData = array
+					(
+						'url' => $strUrl,
+						'content' => $this->strBuffer,
+						'title' => $objPage->pageTitle ?: $objPage->title,
+						'protected' => ($objPage->protected ? '1' : ''),
+						'groups' => $objPage->groups,
+						'pid' => $objPage->id,
+						'language' => $objPage->language
+					);
+
+					\Search::indexPage($arrData);
+				}
 			}
 		}