+++ system/modules/core/classes/DataContainer.php
@@ -610,20 +610,18 @@ class DataContainer extends \Backend
 			$title = is_array($v['label']) ? $v['label'][1] : $v['label'];
 			$attributes = ($v['attributes'] != '') ? ' ' . ltrim($v['attributes']) : '';
 
-			// icon
+			// Custom icon (see #5541)
 			if ($v['icon'])
 			{
 				$v['class'] = trim($v['class'] . ' header_icon');
-				$style = 'background-image:url(' . $v['icon'] . ');';
 
-				if (strpos($attributes, 'style="') !== false)
+				// Add the theme path if only the file name is given
+				if (strpos($v['icon'], '/') === false)
 				{
-					$attributes = preg_replace('/(style=")([^"]*")/','$1' . $style . '$2', $attributes);
+					$v['icon'] = 'system/themes/' . \Backend::getTheme() . '/images/' . $v['icon'];
 				}
-				else
-				{
-					$attributes .= ' style="' . $style . '"';
-				}
+
+				$attributes = sprintf('style="background-image:url(\'%s%s\')"', TL_ASSETS_URL, $v['icon']) . $attributes;
 			}
 
 			if ($label == '')