+++ system/docs/CHANGELOG.md
@@ -4,6 +4,10 @@ Contao Open Source CMS Changelog
 Version 3.1.RC1 (2013-04-XX)
 ----------------------------
 
+### Changed
+The tablesort script can now be added as `moo_` or `j_` template in the page
+layout, so the library can be replaced if needed.
+
 ### Updated
 Updated Swipe.js to version 2.0 and reworked the slider implementation.