+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.0.7 (2013-XX-XX)
 --------------------------
 
 ### Fixed
+Remove script tags when printing an article as PDF (see #5626).
+
+### Fixed
 Purge the internal cache when disabling the Contao safe mode (see #5579).
 
 ### Fixed